requirements.txt
django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
djongo==1.3.6
pymongo==3.12.3
python-dotenv==1.0.0
gunicorn==21.2.0
dnspython==2.4.2
PyJWT==2.8.0
sqlparse==0.2.4 